--- src/main/kotlin/dev/minuk/otelcommander/domain/models/agent/AgentConfigMap.kt
@@ -1,6 +1,0 @@
-package dev.minuk.otelcommander.domain.models.agent
-
-data class AgentConfigMap(
-    val body: String,
-    val contentType: String,
-)
